backend/requirement.txt
astroid
Click
ecdsa
Flask
Flask-SQLAlchemy
future
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
pylint
pycryptodomex
Python-jose
six
typed-ast
Werkzeug
wrapt
Flask-Cors